Fix manage.py location

+++ get-from-pootle.sh
@@ -7,7 +7,8 @@ langs=$*
 user=pootlesync
 server=pootle.locamotion.org
 local_copy=.pootle_tmp
-manage_command="/var/www/sites/$instance/src/pootle/manage.py"
+phaselist=
+manage_command="/var/www/sites/$instance/src/manage.py"
 manage_py_verbosity=2
 precommand=". /var/www/sites/mozilla/env/bin/activate;"
 
@@ -50,24 +51,32 @@ for lang in $langs
 do
 	echo "Language: $lang"
 	cd $lang
-	for phase in $(ls)
-	do
-		if [ -d $phase ]; then
-	        	cd $phase
-	        	echo "Phase: $phase"
-	        	for po in $(find . -name "*.po")
-	        	do
-	        	        mkdir -p $svndir/$lang/$(dirname $po)
-	        	        mv $po $svndir/$lang/$po
-	        	done
-	        	cd ..
-		else
-			mv $phase $svndir/$lang
-		fi
-	done
+	if [ -f "$phaselist" ]; then
+		for phase in $(ls)
+		do
+			if [ -d $phase ]; then
+		        	cd $phase
+		        	echo "Phase: $phase"
+		        	for po in $(find . -name "*.po")
+		        	do
+		        	        mkdir -p $svndir/$lang/$(dirname $po)
+		        	        mv $po $svndir/$lang/$po
+		        	done
+		        	cd ..
+			else
+				mv $phase $svndir/$lang
+			fi
+		done
+	else
+	       	for po in $(find . -name "*.po")
+	       	do
+	       	        mkdir -p $svndir/$lang/$(dirname $po)
+	       	        mv $po $svndir/$lang/$po
+	       	done
+	fi
 	cd ..
 done
 cd ../..
 
 # Cleanup
-./cleanup-msgcat.sh $langs
+#./cleanup-msgcat.sh $langs
